chore: replace unnecessary dependency "is-string-and-not-blank"
This package must be a joke, it contains literally three lines of code
and another dependency - is-string-blank.

The is-string-blank package is 6 years old and the source repository is
gone. Moreover, it can be replaced with just a single line of code -
it's used to check just a very short strings so the performance
difference is negligible, if it still exists.